+++ tsdb/compact.go
@@ -90,7 +90,7 @@ type LeveledCompactor struct {
 	postingsEncoder             index.PostingsEncoder
 	postingsDecoderFactory      PostingsDecoderFactory
 	enableOverlappingCompaction bool
-	cacheAllSymbols             bool
+	CacheOnlySeriesSymbols      bool
 }
 
 type CompactorMetrics struct {
@@ -175,8 +175,8 @@ type LeveledCompactorOptions struct {
 	Metrics *CompactorMetrics
 	// UseUncachedIO allows bypassing the page cache when appropriate.
 	UseUncachedIO bool
-	// CacheAllSymbols enables caching of all TSDB symbols for compaction.
-	CacheAllSymbols bool
+	// CacheOnlySeriesSymbols enables caching of TSDB symbols only when adding series.
+	CacheOnlySeriesSymbols bool
 }
 
 type PostingsDecoderFactory func(meta *BlockMeta) index.PostingsDecoder
@@ -186,23 +186,21 @@ func DefaultPostingsDecoderFactory(_ *BlockMeta) index.PostingsDecoder {
 }
 
 // NewLeveledCompactorWithChunkSize returns a new LeveledCompactor with a certain max block segment chunk size.
-// It's the same as calling NewLeveledCompactorWithOptions with maxBlockChunkSegmentSize, mergeFunc, enabled overlapping compaction, and caching of all symbols during compaction.
+// It's the same as calling NewLeveledCompactorWithOptions with maxBlockChunkSegmentSize, mergeFunc, and enabled overlapping compaction.
 func NewLeveledCompactorWithChunkSize(ctx context.Context, r prometheus.Registerer, l *slog.Logger, ranges []int64, pool chunkenc.Pool, maxBlockChunkSegmentSize int64, mergeFunc storage.VerticalChunkSeriesMergeFunc) (*LeveledCompactor, error) {
 	return NewLeveledCompactorWithOptions(ctx, r, l, ranges, pool, LeveledCompactorOptions{
 		MaxBlockChunkSegmentSize:    maxBlockChunkSegmentSize,
 		MergeFunc:                   mergeFunc,
 		EnableOverlappingCompaction: true,
-		CacheAllSymbols:             true,
 	})
 }
 
 // NewLeveledCompactor returns a new LeveledCompactor.
-// It's the same as calling NewLeveledCompactorWithOptions with mergeFunc, enabled overlapping compaction, and caching of all symbols during compaction.
+// It's the same as calling NewLeveledCompactorWithOptions with mergeFunc, and enabled overlapping compaction.
 func NewLeveledCompactor(ctx context.Context, r prometheus.Registerer, l *slog.Logger, ranges []int64, pool chunkenc.Pool, mergeFunc storage.VerticalChunkSeriesMergeFunc) (*LeveledCompactor, error) {
 	return NewLeveledCompactorWithOptions(ctx, r, l, ranges, pool, LeveledCompactorOptions{
 		MergeFunc:                   mergeFunc,
 		EnableOverlappingCompaction: true,
-		CacheAllSymbols:             true,
 	})
 }
 
@@ -243,7 +241,7 @@ func NewLeveledCompactorWithOptions(ctx context.Context, r prometheus.Registerer
 		postingsEncoder:             pe,
 		postingsDecoderFactory:      opts.PD,
 		enableOverlappingCompaction: opts.EnableOverlappingCompaction,
-		cacheAllSymbols:             opts.CacheAllSymbols,
+		CacheOnlySeriesSymbols:      opts.CacheOnlySeriesSymbols,
 	}, nil
 }
 
@@ -686,7 +684,7 @@ func (c *LeveledCompactor) write(dest string, meta *BlockMeta, blockPopulator Bl
 		}
 	}
 
-	indexw, err := index.NewWriterWithEncoder(c.ctx, filepath.Join(tmp, indexFilename), c.postingsEncoder, c.cacheAllSymbols)
+	indexw, err := index.NewWriterWithEncoder(c.ctx, filepath.Join(tmp, indexFilename), c.postingsEncoder, c.CacheOnlySeriesSymbols)
 	if err != nil {
 		return fmt.Errorf("open index writer: %w", err)
 	}