+++ tsdb/db.go
@@ -90,7 +90,7 @@ func DefaultOptions() *Options {
 		EnableOverlappingCompaction: true,
 		EnableSharding:              false,
 		EnableDelayedCompaction:     false,
-		CacheAllSymbols:             true,
+		CacheOnlySeriesSymbols:      false,
 		CompactionDelayMaxPercent:   DefaultCompactionDelayMaxPercent,
 		CompactionDelay:             time.Duration(0),
 		PostingsDecoderFactory:      DefaultPostingsDecoderFactory,
@@ -208,8 +208,8 @@ type Options struct {
 	// CompactionDelayMaxPercent is the upper limit for CompactionDelay, specified as a percentage of the head chunk range.
 	CompactionDelayMaxPercent int
 
-	// CacheAllSymbols enables caching of all TSDB symbols for compaction.
-	CacheAllSymbols bool
+	// CacheOnlySeriesSymbols enables caching of TSDB symbols only when adding series, during compaction.
+	CacheOnlySeriesSymbols bool
 
 	// NewCompactorFunc is a function that returns a TSDB compactor.
 	NewCompactorFunc NewCompactorFunc
@@ -433,16 +433,16 @@ var ErrClosed = errors.New("db already closed")
 // Current implementation doesn't support concurrency so
 // all API calls should happen in the same go routine.
 type DBReadOnly struct {
-	logger          *slog.Logger
-	dir             string
-	sandboxDir      string
-	cacheAllSymbols bool
-	closers         []io.Closer
-	closed          chan struct{}
+	logger                 *slog.Logger
+	dir                    string
+	sandboxDir             string
+	CacheOnlySeriesSymbols bool
+	closers                []io.Closer
+	closed                 chan struct{}
 }
 
 // OpenDBReadOnly opens DB in the given directory for read only operations.
-func OpenDBReadOnly(dir, sandboxDirRoot string, cacheAllSymbols bool, l *slog.Logger) (*DBReadOnly, error) {
+func OpenDBReadOnly(dir, sandboxDirRoot string, cacheOnlySeriesSymbols bool, l *slog.Logger) (*DBReadOnly, error) {
 	if _, err := os.Stat(dir); err != nil {
 		return nil, fmt.Errorf("opening the db dir: %w", err)
 	}
@@ -460,11 +460,11 @@ func OpenDBReadOnly(dir, sandboxDirRoot string, cacheAllSymbols bool, l *slog.Lo
 	}
 
 	return &DBReadOnly{
-		logger:          l,
-		dir:             dir,
-		cacheAllSymbols: cacheAllSymbols,
-		sandboxDir:      sandboxDir,
-		closed:          make(chan struct{}),
+		logger:                 l,
+		dir:                    dir,
+		CacheOnlySeriesSymbols: cacheOnlySeriesSymbols,
+		sandboxDir:             sandboxDir,
+		closed:                 make(chan struct{}),
 	}, nil
 }
 
@@ -522,7 +522,7 @@ func (db *DBReadOnly) FlushWAL(dir string) (returnErr error) {
 		chunkenc.NewPool(),
 		LeveledCompactorOptions{
 			EnableOverlappingCompaction: true,
-			CacheAllSymbols:             db.cacheAllSymbols,
+			CacheOnlySeriesSymbols:      db.CacheOnlySeriesSymbols,
 		},
 	)
 	if err != nil {
@@ -915,7 +915,7 @@ func open(dir string, l *slog.Logger, r prometheus.Registerer, opts *Options, rn
 		db.compactor, err = NewLeveledCompactorWithOptions(ctx, r, l, rngs, db.chunkPool, LeveledCompactorOptions{
 			MaxBlockChunkSegmentSize:    opts.MaxBlockChunkSegmentSize,
 			EnableOverlappingCompaction: opts.EnableOverlappingCompaction,
-			CacheAllSymbols:             opts.CacheAllSymbols,
+			CacheOnlySeriesSymbols:      opts.CacheOnlySeriesSymbols,
 			PD:                          opts.PostingsDecoderFactory,
 			UseUncachedIO:               opts.UseUncachedIO,
 		})